+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.29
 ### is.codion.swing.common.ui
 - FilterTable.ScrollToAdded bug fixed, now stops scrolling after the first row.
+- FilterTable.Builder.surrendersFocusOnKeystroke() added.
 ### is.codion.framework.db
 - EntityConnectionProvider.connection() inferred return type removed.
 - AbstractEntityConnectionProvider.validateConnection() replaced with validConnection(), connection() no longer final.